refactor: rename q_control module to queue_control and update related references
