+++ orchestrator/src/types/mod.rs
@@ -3,6 +3,6 @@ pub mod constant;
 pub mod error;
 pub mod jobs;
 pub mod params;
-pub mod q_control;
+pub mod queue_control;
 pub mod queue;
 pub mod worker;
